Prevent email enumeration by wrapping email send in a promise
